pyproject.toml
[project]
name = "printhub-zpl-ii"
version = "0.1.0"
description = ""
authors = [
    {name = "Nathaniel Hartmann",email = "hartmannlight@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "win32printing (>=0.1.3,<0.2.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
